tech(roll) - add tests on rolls
