+++ .babelrc
@@ -1,0 +1,10 @@
+{
+    "presets": [
+        "@babel/preset-env"
+    ],
+    "plugins": [
+        [
+            "@babel/transform-runtime"
+        ]
+    ]
+}
