+++ jest.config.json
@@ -1,0 +1,10 @@
+{
+  "collectCoverage": true,
+  "moduleFileExtensions": [
+    "js"
+  ],
+  "transform": {
+    "^.+\\.js$": "babel-jest"
+  },
+  "testRegex": "((\\.|/*.)(spec))\\.js?$"
+}
